+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
 - Updated demo dependencies
 - LokiJS is now a dependency of WatermelonDB (although it's only required for use on the web)
 - [Android] fix compile error for kotlin 1.3+ 
+- [hooks] new `useDatabase` hook for consuming the Database Context
+- [TypeScript] new type declarations
 
 ## 0.10.1 - 2019-02-12